- changed types of parameters to avoid unnecessary limitation in data length - change the return type for the byte reader (BYTE instead of WORD)

+++ src/utils/BitsStreams.h
@@ -99,7 +99,7 @@ public:
      * usable. Refer to getRemainingBits() to know how many bits are usable
      * @return the bits
      */
-    quint16 getBits();
+    quint8 getBits();
 };
 
 /**
